use dialog before running action command

+++ systemact.sh
@@ -94,15 +94,59 @@ case "$1" in
         $lockcmd
         ;;
     logout)
-        $logoutcmd
+        act_image="system-log-out"
+        text_title="Logout"
+        text_msg="The system will log out automatically in 60 seconds."
+        title_var="Logout"
+        btn_act="Logout now"
+        success_msg="logging out"
+        cancel_msg="logout cancelled"
+        yad_confirm_dialog
+        ret=$?
+        if [ "$ret" -eq 0 ]; then
+            $logoutcmd
+        fi
         ;;
     shutdown|poweroff)
-        $ctl poweroff
+        act_image="system-shutdown"
+        text_title="Power Off"
+        text_msg="The system will shutdown automatically in 60 seconds."
+        title_var="Shutdown"
+        btn_act="Shutdown now"
+        success_msg="shutting down"
+        cancel_msg="cancelled shutdown"
+        yad_confirm_dialog
+        ret=$?
+        if [ "$ret" -eq 0 ]; then
+            $ctl poweroff
+        fi
         ;;
     reboot|restart)
-        $ctl reboot
+        act_image="system-reboot"
+        text_title="Restart"
+        text_msg="The system will reboot automatically in 60 seconds."
+        title_var="Reboot"
+        btn_act="Rebooot now"
+        success_msg="rebooting"
+        cancel_msg="cancelled reboot"
+        yad_confirm_dialog
+        ret=$?
+        if [ "$ret" -eq 0 ]; then
+            $ctl reboot
+        fi
         ;;
     suspend|sleep)
-        $ctl "$suspend_method"
+        act_image="system-suspend"
+        text_title="Sleep"
+        text_msg="The system will suspend automatically in 60 seconds."
+        title_var="Suspend"
+        btn_act="Suspend now"
+        success_msg="suspending now"
+        cancel_msg="cancelled suspend"
+        yad_confirm_dialog
+        ret=$?
+        if [ "$ret" -eq 0 ]; then
+            $ctl "$suspend_method"
+        fi
         ;;
 esac
